test_compute_shader_uniform
